Setup the bases to use a Port to interact tailwind cli
The current binary is for mac os x64, and it is a custom binary

We need to save our custom binaries and fetch depending on the OS.
Similar to how tailwind wrapper works, check if a PR to that project is
welcome.

The idea is generate custom tailwind_cli binary with custom plugins, in
our case daisyui plugin is installed.